requirements.txt
ytmusicapi
spotipy
kivy
kivymd
python-dotenv
